Add accelerator for KeySequences
